Refactor subscription confirmation and unsubscription handlers to use URL parameters
Changes:
- Updated the ConfirmSubscription and Unsubscribe methods to retrieve the token from URL parameters instead of the request body, improving API usability.
- Adjusted the routes in the API Gateway to reflect the new GET method for confirming and unsubscribing, enhancing clarity in the API design.
- Introduced a new httpclient package for managing HTTP client configurations, improving code organization and reusability.

This commit enhances the subscription service's API by streamlining token handling and improving the overall structure of the HTTP client.
