+++ internal/services/api-gateway/internal/proto/weather.proto
